pipeline_old/requirements.txt
google-cloud-storage
google-cloud-bigquery
pandas
pyyaml==6.0